moved progress bars do component
